:sparkles: Feat: 검색창 MainActivity에서 SearchFragment로 이동
