+++ app/src/main/java/com/example/readme/ui/MainActivity.kt
@@ -1,11 +1,7 @@
 package com.example.readme.ui
 
 import android.os.Bundle
-import android.text.Editable
-import android.text.TextWatcher
-import android.view.KeyEvent
 import android.view.View
-import android.view.inputmethod.InputMethodManager
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
@@ -14,9 +10,7 @@ import com.example.readme.databinding.ActivityMainBinding
 import com.example.readme.ui.community.CommunityFragment
 import com.example.readme.ui.home.HomeFragment
 import com.example.readme.ui.mypage.MyPageFragment
-import com.example.readme.ui.search.BookSearchPreviewFragment
 import com.example.readme.ui.search.SearchFragment
-import com.example.readme.ui.search.SearchResultFragment
 
 class MainActivity : AppCompatActivity() {
 
@@ -27,7 +21,6 @@ class MainActivity : AppCompatActivity() {
         binding = DataBindingUtil.setContentView(this, R.layout.activity_main)
         ShowInit()
         setupBottomNavigationView()
-        setupSearchView()
         setSupportActionBar(binding.toolbar)
         // Disable displaying the title in the Toolbar
         supportActionBar?.setDisplayShowTitleEnabled(false)
@@ -66,70 +59,6 @@ class MainActivity : AppCompatActivity() {
         binding.bottomNavigationView.selectedItemId = R.id.navigation_home
     }
 
-    private fun setupSearchView() {
-        // 검색 버튼 클릭 시 SearchResultFragment로 이동
-        binding.searchButton.setOnClickListener {
-            hideKeyboard()
-            val searchResultFragment = SearchResultFragment()
-            val bundle = Bundle()
-            bundle.putString("keyword", binding.searchEditText.text.toString())
-            searchResultFragment.arguments = bundle
-            setFragment(searchResultFragment)
-        }
-
-        // EditText 검색 버튼 클릭 시 SearchResultFragment로 이동
-        binding.searchEditText.setOnKeyListener { v, keyCode, event ->
-            if (keyCode == KeyEvent.KEYCODE_ENTER && event.action == KeyEvent.ACTION_UP && binding.searchEditText.text.isNotEmpty()) {
-                hideKeyboard()
-                val searchResultFragment = SearchResultFragment()
-                val bundle = Bundle()
-                bundle.putString("keyword", binding.searchEditText.text.toString())
-                searchResultFragment.arguments = bundle
-                setFragment(searchResultFragment)
-                true
-            } else if (keyCode == KeyEvent.KEYCODE_ENTER && event.action == KeyEvent.ACTION_UP) {
-                hideKeyboard()
-                true
-            } else {
-                false
-            }
-        }
-
-        // EditText 텍스트 변경 시 책 미리보기 검색
-        binding.searchEditText.addTextChangedListener(object : TextWatcher {
-            override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {}
-
-            override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {}
-
-            override fun afterTextChanged(s: Editable?) {
-                val keyword = binding.searchEditText.text.toString()
-                if (keyword.isEmpty()) {
-                    val searchFragment = SearchFragment()
-                    setFragment(searchFragment)
-                } else {
-                    // BookSearchPreviewFragment가 이미 로드되어 있는지 확인
-                    val currentFragment = supportFragmentManager.findFragmentById(R.id.nav_host_fragment)
-                    if (currentFragment is BookSearchPreviewFragment) {
-                        currentFragment.viewModel.onQueryTextChange(keyword)
-                    } else {
-                        // BookSearchPreviewFragment로 전환
-                        val searchBookPreviewFragment = BookSearchPreviewFragment()
-                        val bundle = Bundle()
-                        bundle.putString("keyword", keyword)
-                        searchBookPreviewFragment.arguments = bundle
-                        setFragment(searchBookPreviewFragment)
-                    }
-                }
-            }
-        })
-    }
-
-    private fun hideKeyboard() {
-        val imm = getSystemService(INPUT_METHOD_SERVICE) as InputMethodManager
-        imm.hideSoftInputFromWindow(binding.searchEditText.windowToken, 0)
-        binding.searchEditText.clearFocus() // 포커스 해제
-    }
-
 
     private fun setFragment(fragment: Fragment) {
         supportFragmentManager.beginTransaction()
@@ -138,27 +67,27 @@ class MainActivity : AppCompatActivity() {
     }
 
     fun ShowInit(){
+        binding.toolbar.visibility = View.VISIBLE
         binding.mainLogo.visibility = View.VISIBLE
         binding.btnFilter.visibility = View.VISIBLE
         binding.btnSetting.visibility = View.GONE
         binding.btnBack.visibility = View.GONE
-        binding.searchView.visibility = View.GONE
     }
 
     fun ShowHome(){
+        binding.toolbar.visibility = View.VISIBLE
         binding.mainLogo.visibility = View.VISIBLE
         binding.btnFilter.visibility = View.VISIBLE
         binding.btnSetting.visibility = View.GONE
         binding.btnBack.visibility = View.GONE
-        binding.searchView.visibility = View.GONE
     }
 
     fun ShowSearch(){
+        binding.toolbar.visibility = View.GONE
         binding.mainLogo.visibility = View.GONE
         binding.btnFilter.visibility = View.GONE
         binding.btnSetting.visibility = View.GONE
         binding.btnBack.visibility = View.GONE
-        binding.searchView.visibility = View.VISIBLE
     }
 
     fun ShowCommunity() {
@@ -167,14 +96,13 @@ class MainActivity : AppCompatActivity() {
         binding.btnFilter.visibility = View.GONE
         binding.btnSetting.visibility = View.GONE
         binding.btnBack.visibility = View.GONE
-        binding.searchView.visibility = View.GONE
     }
 
     fun ShowMyPage(){
+        binding.toolbar.visibility = View.VISIBLE
         binding.mainLogo.visibility = View.VISIBLE
         binding.btnFilter.visibility = View.GONE
         binding.btnSetting.visibility = View.VISIBLE
         binding.btnBack.visibility = View.GONE
-        binding.searchView.visibility = View.GONE
     }
 }
