+++ app/src/main/res/layout/fragment_recent_search.xml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".ui.search.RecentSearchFragment"
+    android:paddingTop="10dp">
+
+    <data>
+        <variable
+            name="viewModel"
+            type="com.example.readme.ui.search.RecentSearchViewModel" />
+    </data>
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/search_section"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:paddingHorizontal="20dp">
+
+        <!-- Recent Searches Title -->
+        <TextView
+            android:id="@+id/recent_searches_title"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="최근 검색"
+            android:textSize="14sp"
+            android:visibility="visible"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.0"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <!-- Recent Searches List -->
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/recent_searches_recycler_view"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:layout_marginVertical="10dp"
+            android:contentDescription="최근 검색된 항목 목록"
+            app:layout_constraintTop_toBottomOf="@id/recent_searches_title"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            tools:listitem="@layout/recent_search_book" />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
